+++ pkg/user/user.go
@@ -75,7 +75,7 @@ func CreateUser(usr *userv3.User) error {
 	if err := json.Unmarshal([]byte(resp), &ur); err != nil {
 		return err
 	}
-	fmt.Println("Recovery URL: ", ur.GetSpec().GetRecoveryUrl())
+	fmt.Println("User registered. Use the following default pass phrase: ", ur.GetSpec().GetPassword())
 	return nil
 }